Fix deterministic AArch64 crash in 64-bit context allocation and lookup cache
Fix three 32-bit assumptions that caused a half-allocated context and a
truncated lookup-cache address on the 64-bit AArch64 (macOS/arm64) build,
which crashed on every run.

1) block_prims_gen.cpp (allocateContext_var): the fast and slow paths scaled
   the context byte size by the smi Tag_Size only (nofVars*4), under-allocating
   a context object by 2x on 64-bit (oopSize=8). For a block with nofVars
   variables this left the closing object fields outside the allocation, so
   the born.limit test passed but the frame ended up half-initialized. On
   AArch64, shll(edx,1) after loading the smi length so the byte count is
   nofVars*oopSize (= nofVars*8). x86 (oopSize==4) is unchanged since the smi
   value already equals the byte count.

2) lookupCache.cpp/.hpp: primary_cache_address()/secondary_cache_address()
   returned int, truncating the heap address to 32 bits on a 64-bit heap,
   corrupting the constant addresses emitted into generated code. Return
   intptr_t.

3) interpreter.cpp (copy_params_into_context): the destination register was
   advanced by 4 bytes instead of oopSize, so on 64-bit each argument was
   written 4 bytes before the previous slot. Advanced by oopSize.

+++ vm/lookup/lookupCache.cpp
@@ -152,10 +152,10 @@ public:
 static cacheElement primary[primary_cache_size];
 static cacheElement secondary[secondary_cache_size];
 
-int lookupCache::primary_cache_address() {
+intptr_t lookupCache::primary_cache_address() {
   return intptr_t(&primary[0]);
 }
-int lookupCache::secondary_cache_address() {
+intptr_t lookupCache::secondary_cache_address() {
   return intptr_t(&secondary[0]);
 }
